requirements.txt
locust
pyyaml
beautifulsoup4